Added column 'attempts'
Shows the amount of attempts (receptions for wr, carries for rb, and passing attemtps for qb) before each success (touchdown).
I do this by using a dict 'att'. I use it for each method and clear it before it begins.

Next thing to normalize would be time. I don't know if it will happen, but it may.